pyproject.toml
[project]
name = "fast-geo-coverage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.13.3",
    "cryptography>=43.0.3",
    "fastapi[standard]>=0.115.2",
    "geoalchemy2>=0.15.2",
    "pydantic-settings>=2.6.0",
    "pymysql>=1.1.1",
    "shapely>=2.0.6",
    "sqlmodel>=0.0.22",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.7.0",
]

[tool.ruff]
line-length = 79
exclude = ["alembic/versions", "migrations/", "venv/", "logs/"]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[tool.ruff.format]
quote-style = "single"
indent-style = "tab"
docstring-code-format = true
